refactor: View 및 ViewResolver 클래스 패키지 이동
