+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/resolver/ViewResolver.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.view.resolver;
+package nextstep.mvc.tobe.resolver;
 
 import nextstep.mvc.tobe.view.View;
